% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbvnorm.R
\name{rbvnorm}
\alias{rbvnorm}
\title{rbvnorm}
\usage{
rbvnorm(iter, init = 0, mu, sigma)
}
\arguments{
\item{iter}{number of samples to draw from the distribution}

\item{init}{initial x2 value for sampling}

\item{mu}{vector of means for x1 and x2}

\item{sigma}{covariance matrix for x1 and x2}
}
\value{
a list object containing the random sample and relevant metrix (iter, mu, sigma)
}
\description{
This function allows the user to produce a random sample of the bivariate normal distribution.
}
\examples{
rbvnorm(iter=10000, init=2, mu=c(3,6), sigma = matrix(c(8,-2,-2,4), nrow=2, byrow=TRUE))
}
