% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyplot.R
\name{shinyplots}
\alias{shinyplots}
\title{shinyplot}
\usage{
shinyplots()
}
\value{
Shiny app with contour plot of gibbs sampling
}
\description{
Calls Gibbs sampler shiny app with descriptive plotting and widgets for
}
\examples{
\dontrun{ shinyplots()}
}
